Sprint 4

+++ src/main/java/com/onkibot/backend/api/CategoryController.java
@@ -3,14 +3,17 @@ package com.onkibot.backend.api;
 import com.onkibot.backend.OnkibotBackendApplication;
 import com.onkibot.backend.database.entities.Category;
 import com.onkibot.backend.database.entities.Course;
+import com.onkibot.backend.database.entities.User;
 import com.onkibot.backend.database.repositories.CategoryRepository;
 import com.onkibot.backend.database.repositories.CourseRepository;
+import com.onkibot.backend.database.repositories.UserRepository;
 import com.onkibot.backend.exceptions.CategoryNotFoundException;
 import com.onkibot.backend.exceptions.CourseNotFoundException;
 import com.onkibot.backend.models.CategoryInputModel;
 import com.onkibot.backend.models.CategoryModel;
 import java.util.Collection;
 import java.util.stream.Collectors;
+import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +25,8 @@ public class CategoryController {
   @Autowired private CategoryRepository categoryRepository;
 
   @Autowired private CourseRepository courseRepository;
+
+  @Autowired private UserRepository userRepository;
 
   @RequestMapping(method = RequestMethod.GET, value = "/{categoryId}")
   CategoryModel get(@PathVariable int courseId, @PathVariable int categoryId) {
@@ -46,8 +51,13 @@ public class CategoryController {
   }
 
   @RequestMapping(method = RequestMethod.DELETE, value = "/{categoryId}")
-  public ResponseEntity<Void> delete(@PathVariable int courseId, @PathVariable int categoryId) {
+  public ResponseEntity<Void> delete(
+      @PathVariable int courseId, @PathVariable int categoryId, HttpSession session) {
+    User user = OnkibotBackendApplication.assertSessionUser(userRepository, session);
     Category category = this.assertCourseCategory(courseId, categoryId);
+    if (!user.getIsInstructor() || !user.isAttending(category.getCourse())) {
+      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
     categoryRepository.delete(category);
     return new ResponseEntity<>(HttpStatus.NO_CONTENT);
   }
